requirements.txt
# Meeting Recap App - Python Dependencies

# Core audio/ML processing
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
openai-whisper>=20230314
faster-whisper>=0.10.0

# HTTP client for API communication
requests>=2.31.0

# FastAPI server dependencies (for local API server)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Optional: for better JSON handling
aiofiles>=23.1.0
